Rename DatasetDetail to Dataset and update file structure
