+++ context/app/static/js/components/Detail/Dataset/index.js
@@ -1,0 +1,3 @@
+import Dataset from './Dataset';
+
+export default Dataset;
