+++ context/app/static/js/components/Detail/Dataset/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import Divider from '@material-ui/core/Divider';
+
+const StyledDivider = styled(Divider)`
+  margin-left: 5px;
+  margin-right: 5px;
+  height: 15px;
+  background-color: #444a65;
+  align-self: center;
+`;
+
+export { StyledDivider };
